split existing fixtures into datastore specific files
