+++ tests/conftest.py
@@ -31,9 +31,12 @@ from fidesops.util.oauth_util import generate_jwe
 from .fixtures.application_fixtures import *
 from .fixtures.integration_fixtures import *
 from .fixtures.mariadb_fixtures import *
+from .fixtures.mongodb_fixtures import *
 from .fixtures.mysql_fixtures import *
 from .fixtures.mssql_fixtures import *
 from .fixtures.postgres_fixtures import *
+from .fixtures.redshift_fixtures import *
+from .fixtures.snowflake_fixtures import *
 
 logger = logging.getLogger(__name__)
