+++ tests/fixtures/application_fixtures.py
@@ -36,10 +36,6 @@ from fidesops.models.privacy_request import (
     ExecutionLogStatus,
 )
 from fidesops.models.storage import StorageConfig, ResponseFormat
-from fidesops.schemas.connection_configuration import (
-    SnowflakeSchema,
-    RedshiftSchema,
-)
 from fidesops.schemas.storage.storage import (
     FileNaming,
     StorageDetails,
@@ -163,170 +159,6 @@ def storage_config_onetrust(db: Session) -> Generator:
     )
     yield storage_config
     storage_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def connection_config(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_postgres_db_1",
-            "connection_type": ConnectionType.postgres,
-            "access": AccessLevel.write,
-            "secrets": integration_secrets["postgres_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def read_connection_config(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_postgres_db_1_read_config",
-            "connection_type": ConnectionType.postgres,
-            "access": AccessLevel.read,
-            "secrets": integration_secrets["postgres_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def connection_config_mysql(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_mysql_db_1",
-            "connection_type": ConnectionType.mysql,
-            "access": AccessLevel.write,
-            "secrets": integration_secrets["mysql_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def connection_config_mssql(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_mssql_db_1",
-            "connection_type": ConnectionType.mssql,
-            "access": AccessLevel.write,
-            "secrets": integration_secrets["mssql_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def connection_config_dry_run(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "postgres",
-            "connection_type": ConnectionType.postgres,
-            "access": AccessLevel.write,
-            "secrets": integration_secrets["postgres_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def mongo_connection_config(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_mongo_db_1",
-            "connection_type": ConnectionType.mongodb,
-            "access": AccessLevel.write,
-            "secrets": integration_secrets["mongo_example"],
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def redshift_connection_config(db: Session) -> Generator:
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_redshift_config",
-            "connection_type": ConnectionType.redshift,
-            "access": AccessLevel.write,
-        },
-    )
-    uri = integration_config.get("redshift", {}).get("external_uri") or os.environ.get(
-        "REDSHIFT_TEST_URI"
-    )
-    db_schema = integration_config.get("redshift", {}).get(
-        "db_schema"
-    ) or os.environ.get("REDSHIFT_TEST_DB_SCHEMA")
-    if uri and db_schema:
-        schema = RedshiftSchema(url=uri, db_schema=db_schema)
-        connection_config.secrets = schema.dict()
-        connection_config.save(db=db)
-
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def snowflake_connection_config_without_secrets(
-    db: Session,
-) -> Generator:
-    """
-    Returns a Snowflake ConnectionConfig without secrets
-    attached that is safe to usein any tests.
-    """
-    connection_config = ConnectionConfig.create(
-        db=db,
-        data={
-            "name": str(uuid4()),
-            "key": "my_snowflake_config",
-            "connection_type": ConnectionType.snowflake,
-            "access": AccessLevel.write,
-        },
-    )
-    yield connection_config
-    connection_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def snowflake_connection_config(
-    db: Session,
-    integration_config: Dict[str, str],
-    snowflake_connection_config_without_secrets: ConnectionConfig,
-) -> Generator:
-    """
-    Returns a Snowflake ConectionConfig with secrets attached if secrets are present
-    in the configuration.
-    """
-    snowflake_connection_config = snowflake_connection_config_without_secrets
-    uri = integration_config.get("snowflake", {}).get("external_uri") or os.environ.get(
-        "SNOWFLAKE_TEST_URI"
-    )
-    if uri is not None:
-        schema = SnowflakeSchema(url=uri)
-        snowflake_connection_config.secrets = schema.dict()
-        snowflake_connection_config.save(db=db)
-    yield snowflake_connection_config
 
 
 @pytest.fixture(scope="function")
@@ -774,86 +606,10 @@ def failed_privacy_request(db: Session, policy: Policy) -> PrivacyRequest:
 
 
 @pytest.fixture(scope="function")
-def postgres_execution_log(
-    db: Session, privacy_request: PrivacyRequest
-) -> ExecutionLog:
-    el = ExecutionLog.create(
-        db=db,
-        data={
-            "dataset_name": "my-postgres-db",
-            "collection_name": "user",
-            "fields_affected": [
-                {
-                    "path": "my-postgres-db:user:email",
-                    "field_name": "email",
-                    "data_categories": ["user.provided.identifiable.contact.email"],
-                }
-            ],
-            "action_type": ActionType.access,
-            "status": ExecutionLogStatus.pending,
-            "privacy_request_id": privacy_request.id,
-        },
-    )
-    yield el
-    el.delete(db)
-
-
-@pytest.fixture(scope="function")
-def second_postgres_execution_log(
-    db: Session, privacy_request: PrivacyRequest
-) -> ExecutionLog:
-    el = ExecutionLog.create(
-        db=db,
-        data={
-            "dataset_name": "my-postgres-db",
-            "collection_name": "address",
-            "fields_affected": [
-                {
-                    "path": "my-postgres-db:address:street",
-                    "field_name": "street",
-                    "data_categories": ["user.provided.identifiable.contact.street"],
-                },
-                {
-                    "path": "my-postgres-db:address:city",
-                    "field_name": "city",
-                    "data_categories": ["user.provided.identifiable.contact.city"],
-                },
-            ],
-            "action_type": ActionType.access,
-            "status": ExecutionLogStatus.error,
-            "privacy_request_id": privacy_request.id,
-            "message": "Database timed out.",
-        },
-    )
-    yield el
-    el.delete(db)
-
-
-@pytest.fixture(scope="function")
-def mongo_execution_log(db: Session, privacy_request: PrivacyRequest) -> ExecutionLog:
-    el = ExecutionLog.create(
-        db=db,
-        data={
-            "dataset_name": "my-mongo-db",
-            "collection_name": "orders",
-            "fields_affected": [
-                {
-                    "path": "my-mongo-db:orders:name",
-                    "field_name": "name",
-                    "data_categories": ["user.provided.identifiable.contact.name"],
-                }
-            ],
-            "action_type": ActionType.access,
-            "status": ExecutionLogStatus.in_processing,
-            "privacy_request_id": privacy_request.id,
-        },
-    )
-    yield el
-    el.delete(db)
-
-
-@pytest.fixture(scope="function")
-def dataset_config(connection_config: ConnectionConfig, db: Session) -> Generator:
+def dataset_config(
+    connection_config: ConnectionConfig,
+    db: Session,
+) -> Generator:
     dataset_config = DatasetConfig.create(
         db=db,
         data={
@@ -865,46 +621,6 @@ def dataset_config(connection_config: ConnectionConfig, db: Session) -> Generato
                 "description": "Example Postgres dataset created in test fixtures",
                 "dataset_type": "PostgreSQL",
                 "location": "postgres_example.test",
-                "collections": [
-                    {
-                        "name": "subscriptions",
-                        "fields": [
-                            {
-                                "name": "id",
-                                "data_categories": ["system.operations"],
-                            },
-                            {
-                                "name": "email",
-                                "data_categories": [
-                                    "user.provided.identifiable.contact.email"
-                                ],
-                                "fidesops_meta": {
-                                    "identity": "email",
-                                },
-                            },
-                        ],
-                    },
-                ],
-            },
-        },
-    )
-    yield dataset_config
-    dataset_config.delete(db)
-
-
-@pytest.fixture(scope="function")
-def dataset_config_mysql(connection_config: ConnectionConfig, db: Session) -> Generator:
-    dataset_config = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": connection_config.id,
-            "fides_key": "mysql_example_subscriptions_dataset",
-            "dataset": {
-                "fides_key": "mysql_example_subscriptions_dataset",
-                "name": "Mysql Example Subscribers Dataset",
-                "description": "Example Mysql dataset created in test fixtures",
-                "dataset_type": "MySQL",
-                "location": "mysql_example.test",
                 "collections": [
                     {
                         "name": "subscriptions",
@@ -995,135 +711,6 @@ def example_datasets() -> List[Dict]:
     for filename in example_filenames:
         example_datasets += load_dataset(filename)
     return example_datasets
-
-
-@pytest.fixture
-def snowflake_example_test_dataset_config(
-    snowflake_connection_config: ConnectionConfig,
-    db: Session,
-    example_datasets: List[Dict],
-) -> Generator:
-    dataset = example_datasets[2]
-    fides_key = dataset["fides_key"]
-    dataset_config = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": snowflake_connection_config.id,
-            "fides_key": fides_key,
-            "dataset": dataset,
-        },
-    )
-    yield dataset_config
-    dataset_config.delete(db=db)
-
-
-@pytest.fixture
-def redshift_example_test_dataset_config(
-    redshift_connection_config: ConnectionConfig,
-    db: Session,
-    example_datasets: List[Dict],
-) -> Generator:
-    dataset = example_datasets[3]
-    fides_key = dataset["fides_key"]
-    dataset_config = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": redshift_connection_config.id,
-            "fides_key": fides_key,
-            "dataset": dataset,
-        },
-    )
-    yield dataset_config
-    dataset_config.delete(db=db)
-
-
-@pytest.fixture
-def postgres_example_test_dataset_config(
-    connection_config: ConnectionConfig,
-    db: Session,
-    example_datasets: List[Dict],
-) -> Generator:
-    postgres_dataset = example_datasets[0]
-    fides_key = postgres_dataset["fides_key"]
-    connection_config.name = fides_key
-    connection_config.key = fides_key
-    connection_config.save(db=db)
-    dataset = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": connection_config.id,
-            "fides_key": fides_key,
-            "dataset": postgres_dataset,
-        },
-    )
-    yield dataset
-    dataset.delete(db=db)
-
-
-@pytest.fixture
-def mssql_example_test_dataset_config(
-        connection_config_mssql: ConnectionConfig,
-        db: Session,
-        example_datasets: List[Dict],
-) -> Generator:
-    mssql_dataset = example_datasets[4]
-    fides_key = mssql_dataset["fides_key"]
-    connection_config_mssql.name = fides_key
-    connection_config_mssql.key = fides_key
-    connection_config_mssql.save(db=db)
-    dataset = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": connection_config_mssql.id,
-            "fides_key": fides_key,
-            "dataset": mssql_dataset,
-        },
-    )
-    yield dataset
-    dataset.delete(db=db)
-
-
-@pytest.fixture
-def postgres_example_test_dataset_config_read_access(
-    read_connection_config: ConnectionConfig,
-    db: Session,
-    example_datasets: List[Dict],
-) -> Generator:
-    postgres_dataset = example_datasets[0]
-    fides_key = postgres_dataset["fides_key"]
-    dataset = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": read_connection_config.id,
-            "fides_key": fides_key,
-            "dataset": postgres_dataset,
-        },
-    )
-    yield dataset
-    dataset.delete(db=db)
-
-
-@pytest.fixture
-def mysql_example_test_dataset_config(
-    connection_config_mysql: ConnectionConfig,
-    db: Session,
-    example_datasets: List[Dict],
-) -> Generator:
-    mysql_dataset = example_datasets[5]
-    fides_key = mysql_dataset["fides_key"]
-    connection_config_mysql.name = fides_key
-    connection_config_mysql.key = fides_key
-    connection_config_mysql.save(db=db)
-    dataset = DatasetConfig.create(
-        db=db,
-        data={
-            "connection_config_id": connection_config_mysql.id,
-            "fides_key": fides_key,
-            "dataset": mysql_dataset,
-        },
-    )
-    yield dataset
-    dataset.delete(db=db)
 
 
 @pytest.fixture
