+++ tests/fixtures/postgres_fixtures.py
@@ -1,16 +1,166 @@
 import logging
-from os import truncate
 import pytest
+from typing import Dict, Generator, List
+from uuid import uuid4
+
+from sqlalchemy.orm import Session
 
 from fidesops.db.session import get_db_session, get_db_engine
+from fidesops.models.connectionconfig import (
+    ConnectionConfig,
+    AccessLevel,
+    ConnectionType,
+)
+from fidesops.models.datasetconfig import DatasetConfig
+from fidesops.models.policy import ActionType
+from fidesops.models.privacy_request import (
+    ExecutionLog,
+    ExecutionLogStatus,
+    PrivacyRequest,
+)
 from fidesops.service.connectors import PostgreSQLConnector
+
+from .application_fixtures import integration_secrets
 
 logger = logging.getLogger(__name__)
 
 
+@pytest.fixture
+def postgres_example_test_dataset_config(
+    connection_config: ConnectionConfig,
+    db: Session,
+    example_datasets: List[Dict],
+) -> Generator:
+    postgres_dataset = example_datasets[0]
+    fides_key = postgres_dataset["fides_key"]
+    connection_config.name = fides_key
+    connection_config.key = fides_key
+    connection_config.save(db=db)
+    dataset = DatasetConfig.create(
+        db=db,
+        data={
+            "connection_config_id": connection_config.id,
+            "fides_key": fides_key,
+            "dataset": postgres_dataset,
+        },
+    )
+    yield dataset
+    dataset.delete(db=db)
+
+
+@pytest.fixture
+def postgres_example_test_dataset_config_read_access(
+    read_connection_config: ConnectionConfig,
+    db: Session,
+    example_datasets: List[Dict],
+) -> Generator:
+    postgres_dataset = example_datasets[0]
+    fides_key = postgres_dataset["fides_key"]
+    dataset = DatasetConfig.create(
+        db=db,
+        data={
+            "connection_config_id": read_connection_config.id,
+            "fides_key": fides_key,
+            "dataset": postgres_dataset,
+        },
+    )
+    yield dataset
+    dataset.delete(db=db)
+
+
 @pytest.fixture(scope="function")
-def postgres_integration_session_cls(connection_config):
-    example_postgres_uri = PostgreSQLConnector(connection_config).build_uri()
+def postgres_execution_log(
+    db: Session,
+    privacy_request: PrivacyRequest,
+) -> ExecutionLog:
+    el = ExecutionLog.create(
+        db=db,
+        data={
+            "dataset_name": "my-postgres-db",
+            "collection_name": "user",
+            "fields_affected": [
+                {
+                    "path": "my-postgres-db:user:email",
+                    "field_name": "email",
+                    "data_categories": ["user.provided.identifiable.contact.email"],
+                }
+            ],
+            "action_type": ActionType.access,
+            "status": ExecutionLogStatus.pending,
+            "privacy_request_id": privacy_request.id,
+        },
+    )
+    yield el
+    el.delete(db)
+
+
+# TODO: Consolidate these
+@pytest.fixture(scope="function")
+def second_postgres_execution_log(
+    db: Session, privacy_request: PrivacyRequest
+) -> ExecutionLog:
+    el = ExecutionLog.create(
+        db=db,
+        data={
+            "dataset_name": "my-postgres-db",
+            "collection_name": "address",
+            "fields_affected": [
+                {
+                    "path": "my-postgres-db:address:street",
+                    "field_name": "street",
+                    "data_categories": ["user.provided.identifiable.contact.street"],
+                },
+                {
+                    "path": "my-postgres-db:address:city",
+                    "field_name": "city",
+                    "data_categories": ["user.provided.identifiable.contact.city"],
+                },
+            ],
+            "action_type": ActionType.access,
+            "status": ExecutionLogStatus.error,
+            "privacy_request_id": privacy_request.id,
+            "message": "Database timed out.",
+        },
+    )
+    yield el
+    el.delete(db)
+
+
+@pytest.fixture(scope="function")
+def connection_config(db: Session) -> Generator:
+    connection_config = ConnectionConfig.create(
+        db=db,
+        data={
+            "name": str(uuid4()),
+            "key": "my_postgres_db_1",
+            "connection_type": ConnectionType.postgres,
+            "access": AccessLevel.write,
+            "secrets": integration_secrets["postgres_example"],
+        },
+    )
+    yield connection_config
+    connection_config.delete(db)
+
+
+@pytest.fixture(scope="function")
+def read_connection_config(db: Session) -> Generator:
+    connection_config = ConnectionConfig.create(
+        db=db,
+        data={
+            "name": str(uuid4()),
+            "key": "my_postgres_db_1_read_config",
+            "connection_type": ConnectionType.postgres,
+            "access": AccessLevel.read,
+            "secrets": integration_secrets["postgres_example"],
+        },
+    )
+    yield connection_config
+    connection_config.delete(db)
+
+
+@pytest.fixture(scope="function")
+def postgres_integration_session_cls(connection_config_postgres):
+    example_postgres_uri = PostgreSQLConnector(connection_config_postgres).build_uri()
     engine = get_db_engine(database_uri=example_postgres_uri)
     SessionLocal = get_db_session(
         engine=engine,
